Move write() to global_uinputs.py updated joystick_to_mouse.py
